refactor: session test refactored out
